requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
pydantic==1.10.2
requests==2.31.0
pyannote.audio==2.1.1
torch==2.0.1aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.13.1
Flask==3.0.3
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.21.4
HyperPyYAML==1.2.2
idna==3.6
importlib_metadata==8.5.0
importlib_resources==6.3.1
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.11.0
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
optuna==3.6.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
protobuf==5.26.0
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
spectra-torch==0.4.0
speechbrain==1.0.0
SQLAlchemy==2.0.28
starlette==0.40.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.2
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.32.0
Werkzeug==3.0.4
yarl==1.9.4
zipp==3.20.2
